Address issue #218 + extract dependency file unit test to a proper unit test class
