+++ com.ca.lsp.cobol/lsp-service-cobol/src/main/java/com/ca/lsp/cobol/service/CopybookService.java
@@ -19,12 +19,13 @@ import com.broadcom.lsp.domain.cobol.event.api.EventObserver;
 import com.broadcom.lsp.domain.cobol.event.model.RequiredCopybookEvent;
 import org.eclipse.lsp4j.WorkspaceFolder;
 
-import java.io.IOException;
+import java.nio.file.Path;
 import java.util.List;
 
-public interface FileSystemService extends EventObserver<RequiredCopybookEvent> {
-  String getContentByCopybookName(String copybookName, String profile, List<String> datasetList)
-      throws IOException;
+public interface CopybookService extends EventObserver<RequiredCopybookEvent> {
+  Path findCopybook(String fileName);
+
+  Path findCopybook(String filename, String profile, List<String> datasetList);
 
   void setWorkspaceFolders(List<WorkspaceFolder> workspaceFolders);
 }
