+++ com.ca.lsp.cobol/lsp-service-cobol/src/main/java/com/ca/lsp/cobol/service/delegates/dependency/CopybookDependencyService.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2020 Broadcom.
+ *
+ * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ * Broadcom, Inc. - initial API and implementation
+ *
+ */
+
+package com.ca.lsp.cobol.service.delegates.dependency;
+
+import java.nio.file.Path;
+import java.util.List;
+
+public interface CopybookDependencyService {
+  void setWorkspaceFolderPaths(List<Path> workspaceFolderPaths);
+
+  void generateDependencyFile(String cobolFileName);
+
+  void updateDependencyList(Path dependencyFilePath, String requiredCopybookName);
+
+  void addCopybookInDepFile(String requiredCopybookName, String documentUri);
+}
