+++ com.ca.lsp.cobol/lsp-service-cobol/src/test/java/com/ca/lsp/cobol/service/FileSystemServiceTest.java
@@ -24,23 +24,19 @@ import lombok.extern.slf4j.Slf4j;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Collections;
-import java.util.List;
 
 import static junit.framework.TestCase.assertTrue;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 
 @Slf4j
 public class FileSystemServiceTest extends FileSystemConfiguration
     implements EventObserver<RequiredCopybookEvent> {
 
-  private static final String DEP_EXTENSION = ".dep";
-  private FileSystemServiceImpl fileSystemService =
-      (FileSystemServiceImpl) LangServerCtx.getInjector().getInstance(FileSystemService.class);
+  private CopybookServiceImpl fileSystemService =
+      (CopybookServiceImpl) LangServerCtx.getInjector().getInstance(CopybookService.class);
 
   // Activities performed
   // 1 - Create folder structure in temp folder - Create two copybooks in the provided
@@ -58,17 +54,37 @@ public class FileSystemServiceTest extends FileSystemConfiguration
     fileSystemService.setWorkspaceFolders(generateWorkspaceFolder());
   }
 
+  /**
+   * This test verify that for a given copybook present in the workspace folder the
+   * FileSystemService is able to found it
+   */
+  @Test
+  public void findCopybookByNamePositiveTest() {
+    assertNotNull(fileSystemService.findCopybook(CPY_OUTER_NAME_ONLY2));
+  }
+
+  /**
+   * This test verify that asking for a copybook not present in the workspace folder return null to
+   * the client
+   */
+  @Test
+  public void findCopybookByNameNegativeTest() {
+    assertNull(fileSystemService.findCopybook(COPYBOOK_NOT_PRESENT));
+  }
+
   /*
    * This test take in input a copybook name and MF configuration (profile, target dataset list), and retrieve the
    * content of the copybook, null otherwise.
    */
   @Test
   public void getContentByCopybookName() {
-    assertNotNull(
-        fileSystemService.getContentByCopybookName(
+    Path path =
+        fileSystemService.findCopybook(
             CPY_OUTER_NAME_ONLY2,
             (String) configurationSettingsStorable.getProfiles(),
-            configurationSettingsStorable.getPaths()));
+            configurationSettingsStorable.getPaths());
+
+    assertNotNull(fileSystemService.retrieveContentByPath(path));
   }
 
   /**
@@ -78,8 +94,8 @@ public class FileSystemServiceTest extends FileSystemConfiguration
   @Test
   public void getNullWithNotCopybookNotFound() {
     assertNull(
-        fileSystemService.getContentByCopybookName(
-            "ANTHRCPY",
+        fileSystemService.findCopybook(
+            COPYBOOK_NOT_PRESENT,
             (String) configurationSettingsStorable.getProfiles(),
             configurationSettingsStorable.getPaths()));
   }
@@ -98,51 +114,6 @@ public class FileSystemServiceTest extends FileSystemConfiguration
     assertTrue(
         fetchedCopybookEvent.getName().length() > 0
             && fetchedCopybookEvent.getContent().length() > 0);
-  }
-
-  /**
-   * This test verify that the dependency file is created when a new copybook is found by the
-   * preprocessor.
-   */
-  @Test
-  public void depFileGenerationPositiveTest() {
-    Path depFileReference = createDependencyFile();
-    assertTrue(Files.exists(depFileReference));
-  }
-
-  /**
-   * This test verify that an already defined dependency file is updated with a new copybook
-   * required to be downloaded.
-   */
-  @Test
-  public void depFileUpdatePositiveTest() {
-    // create the dep folder with the SOMPROG.dep
-    Path depFileReference = createDependencyFile();
-
-    // check the number of elements before the update
-    int numberOfElements = getNumberOfElementsFromDepFile(depFileReference);
-
-    // update dep file with a new copybook
-    fileSystemService.updateDependencyList(depFileReference, "ANEWCPY2");
-    // read the list again..
-    // assert that number of element change by 1
-    assertEquals(numberOfElements + 1, getNumberOfElementsFromDepFile(depFileReference));
-  }
-
-  /** This test verify that an empty value for the copybook is not added in the deplist. */
-  @Test
-  public void depFileWithoutEmptyCopybookName() {
-    // create the dep folder with the SOMPROG.dep
-    Path depFileReference = createDependencyFile();
-
-    // check the number of elements before the update
-    int numberOfElements = getNumberOfElementsFromDepFile(depFileReference);
-
-    // update dep file with a new copybook
-    fileSystemService.addCopybookInDepFile(EMPTY_COPYBOOK_NAME, DOCUMENT_URI);
-
-    // assert that number of element didn't change
-    assertEquals(numberOfElements, getNumberOfElementsFromDepFile(depFileReference));
   }
 
   /**
@@ -183,34 +154,6 @@ public class FileSystemServiceTest extends FileSystemConfiguration
             CPY_OUTER_NAME_ONLY2,
             (String) configurationSettingsStorable.getProfiles(),
             Collections.singletonList("HLQLF02.DSNAME1")));
-  }
-
-  private int getNumberOfElementsFromDepFile(Path depFileReference) {
-    List<String> depFileReferenceElementList;
-
-    depFileReferenceElementList = getElementListFromDepFile(depFileReference);
-
-    assert depFileReferenceElementList != null;
-    return depFileReferenceElementList.size();
-  }
-
-  private List<String> getElementListFromDepFile(Path depFileReference) {
-    try {
-      return Files.readAllLines(depFileReference);
-    } catch (IOException e) {
-      log.error(e.getLocalizedMessage());
-      return null;
-    }
-  }
-
-  private Path createDependencyFile() {
-    fileSystemService.generateDependencyFile(DEP_FILE_COST_NAME, getDependencyFolder());
-    return Paths.get(
-        getDependencyFolder() + filesystemSeparator() + DEP_FILE_COST_NAME + DEP_EXTENSION);
-  }
-
-  private Path getDependencyFolder() {
-    return depenencyFileFolderPath;
   }
 
   @Override
